string_nconcat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -11,10 +11,10 @@
  */
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
-	int i;
+	unsigned int i;
 
-	int s1len = 0;
-	int s2len = 0;
+	unsigned int s1len = 0;
+	unsigned int s2len = 0;
 	char *output;
 
 	if (s1 == NULL)
@@ -43,7 +43,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 
 		for (i = 0; s1[i] != '\0'; i++)
 			output[i] = s1[i];
-		for (i = 0;  i < (int)n; i++)
+		for (i = 0;  i < n; i++)
 			output[s1len + i] = s2[i];
 		output[s1len + i] = '\0';
 	}
